requirements.txt
pandas
requests
beautifulsoup4
lxml
openai
python-dotenv
tqdm
